requirements.txt
gym==0.10.11
hydra-core==1.3.2
numpy==1.24.2
omegaconf==2.3.0
torch==1.13.0+cu117
